test(unit.tree): recreate test cases in tests/unit
Moved tests to make the implementations cleaner, and to make managing
tests easier. Added rstest, possibly will add proptest.
