+++ src/tree/core/tree.rs
@@ -101,138 +101,6 @@ impl Display for TreeGenotype {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_access() {
-        let mut tree = TreeGenotype {
-            arena: Vec::new(),
-            children: HashMap::new(),
-        };
-        let arena: Vec<String> = ["+", "*", "2", "x", "-1"].iter().map(|w| w.to_string()).collect();
-        let mut children: HashMap<usize, Vec<usize>> = HashMap::new();
-        children.insert(0, vec![1, 4]);
-        children.insert(1, vec![2, 3]);
-
-        for node in &arena {
-            tree.arena_mut().push(node.to_string());
-        }
-
-        for (&key, value) in children.iter() {
-            tree.children_mut().insert(key, value.clone());
-        }
-        
-        assert_eq!(tree.arena, arena);
-        assert_eq!(tree.children, children);
-    }
-
-    #[test]
-    fn test_subtree() {
-        let mut tree = TreeGenotype {
-            arena: vec![
-                "+".to_string(),
-                "*".to_string(),
-                "2".to_string(),
-                "x".to_string(),
-                "-1".to_string(),
-            ],
-            children: HashMap::new(),
-        };
-        tree.children.insert(0, vec![1, 4]);
-        tree.children.insert(1, vec![2, 3]);
-        
-        assert_eq!(tree.subtree(0), 4);
-        assert_eq!(tree.subtree(1), 3);
-        assert_eq!(tree.subtree(2), 2);
-        assert_eq!(tree.subtree(3), 3);
-        assert_eq!(tree.subtree(4), 4);
-    }
-
-    #[test]
-    fn test_construct_children() {
-        let operators: Vec<String> = ["+", "-", "sin", "x", "y", "z"].iter().map(|&w| w.to_string()).collect();
-        let arity = vec![2, 2, 1, 0, 0, 0];
-        let weights = vec![1.0 / 6.0; 6];
-
-        let sampler = OperatorSampler::new(operators, arity, weights);
-        
-        let arena: Vec<String> = ["+", "*", "2", "x", "-1"].iter().map(|w| w.to_string()).collect();
-        let mut children: HashMap<usize, Vec<usize>> = HashMap::new();
-        children.insert(0, vec![1, 4]);
-        children.insert(1, vec![2, 3]);
-        
-        let tree = TreeGenotype::with_arena(arena.clone());
-
-        let mut test_tree = TreeGenotype { arena, children: HashMap::new() };
-        *test_tree.children_mut() = test_tree.construct_children(&sampler);
-
-        assert_eq!(tree.arena, test_tree.arena);
-        assert!(!test_tree.children().is_empty());
-        assert_ne!(tree.children(), test_tree.children());
-    }
-
-    #[test]
-    fn test_construct_children2() {
-        let operators: Vec<String> = ["+", "-", "sin", "x", "y", "z"].iter().map(|&w| w.to_string()).collect();
-        let arity = vec![2, 2, 1, 0, 0, 0];
-        let weights = vec![1.0 / 6.0; 6];
-
-        let sampler = OperatorSampler::new(operators, arity, weights);
-        
-        let arena: Vec<String> = ["-", "-", "-", "y", "sin", "y", "z", "-", "sin", "+", "x", "y", "y"].iter()
-            .map(|w| w.to_string()).collect();
-        let mut children: HashMap<usize, Vec<usize>> = HashMap::new();
-        children.insert(0, vec![1, 7]);
-        children.insert(1, vec![2, 6]);
-        children.insert(2, vec![3, 4]);
-        children.insert(4, vec![5]);
-        children.insert(7, vec![8, 12]);
-        children.insert(8, vec![9]);
-        children.insert(9, vec![10, 11]);
-        
-        let tree = TreeGenotype::new(arena.clone(), children.clone());
-
-        let mut test_tree = TreeGenotype::with_arena(arena);
-        *test_tree.children_mut() = test_tree.construct_children(&sampler);
-        
-        println!("{:?}", test_tree.arena());
-        println!("{:?}", test_tree.children());
-
-        assert_eq!(tree.arena, test_tree.arena);
-        assert!(!test_tree.children().is_empty());
-        assert_eq!(tree.children(), test_tree.children());
-    }
-
-    #[test]
-    fn test_tree_display() {
-        let tree = TreeGenotype::new(Vec::new(), HashMap::new());
-
-        println!("{}", tree);
-
-        let mut tree = TreeGenotype {
-            arena: vec![
-                "+".to_string(),
-                "*".to_string(),
-                "2".to_string(),
-                "x".to_string(),
-                "-1".to_string(),
-            ],
-            children: HashMap::new(),
-        };
-        tree.children.insert(0, vec![1, 4]);
-        tree.children.insert(1, vec![2, 3]);
-
-        let expected_output = "\
-+
-├── *
-│   ├── 2
-│   └── x
-└── -1
-";
-        println!("{}", tree);
-        assert_eq!(format!("{}", tree), expected_output);
-    }
+impl Default for TreeGenotype {
+    fn default() -> Self { return Self { arena: Vec::new(), children: HashMap::new() }; }
 }
-
